requirements.txt
alembic==1.15.1
anyio==4.8.0
asgiref==3.8.1
asyncio==3.4.3
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lz==0.11.1
Mako==1.3.9
MarkupSafe==3.0.2
memoir==0.0.3
mypy==0.991
mypy-extensions==1.0.0
paradigm==0.6.2
pillow==11.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-telegram-bot==21.11.1
reprit==0.9.0
sniffio==1.3.1
SQLAlchemy==2.0.39
tdlib==0.0.1
telegram_text==0.2.0
typing_extensions==4.12.2
Werkzeug==3.1.3
